Don't require userdata.img staying in the same dir with other files.
With this change, we could use a blaze rule to generate a sightly different
data image and then launch an emulator session with the modified data image.
Directboot test needs a image come with preset lock pin and a custom data image
is a easy way to get this done.

PiperOrigin-RevId: 157169469