Default emulator in make_device
Currently there are device targets with no emulator suffix, which are implicitly qemu1, and other device targets with a _qemu2 suffix.

We want to change the default (suffixless) targets to be qemu2, but we first need explicit qemu1 device targets, so that projects which have issues with qemu2 will have a way to go back to qemu1.

So I'm changing how the default device targets are generated. Previously the qemu1 device object had suffix=False, but I'm getting rid of that flag and instead adding a default_emulator_type param in make_device, and a default_emulator param in new_devices.

If the default emulator is not specified, we'll fallback to the "default default", which is currently qemu1, but we can easily change that later (by API level if we want to).

If you want to compare the targets before and after, here are the targets for generic_tablet:
- old: go/paste/5092275881246720
- new: go/paste/6372588452315136
- diff: go/paste/5228227031203840

As you can see, most of the newly added targets have a _qemu suffix. For API 26, which doesn't support qemu1, this also adds suffixless targets which will use qemu2.

PiperOrigin-RevId: 167396030