Add pike as an emulated device.
The Polar M600 represents an edge case because it has a really small screen, so it is an interesting device to test on in scuba tests to make sure we're not breaking the layout on such a small screen.

Got the device specs here: go/cw-device-specs.

TESTED=
Ran a scuba test with this on API 25; seemed to match the device specs.
PiperOrigin-RevId: 169367105