Branch devices with wrong dpi to wrong_dpi_* targets.
This change is a NOP, since we keep the original target names.

Background:
The emulator had a limitation on large display resolution.
As a workaround to support larger display, we reduced the resolution
and dpi by 50%. The latest emulator supports larger displays and we
are going to update the resolution/dpi to the right values, but this
will break all screendiffing tests (aka. scuba tests).
PiperOrigin-RevId: 178995983