Make qemu2 the default emulator for most devices
See this announcement for more info: []

If I added you as a reviewer, please review just for your project's files (except str/thomaswk).

qemu2 was already the default for API 25+. This change makes it the default for all API levels, except for ARM devices and wear devices with API 20-23. I'll follow up with separate CLs later to eliminate those exceptions.

A handful of target had issues on qemu2. See [1] for the failures. I fixed a few trivial issues, but most of the failing tests I just pinned to qemu1 for now. I'll follow up on them later.

[1] []

PiperOrigin-RevId: 179593952