Create a new emulator target with increased memory for SCUBA tests on Wear.
Wear devices usually have limited memory. Furthermore, because of small screens, cards rendered on Wear devices often wind up being several screens in height. This means that the size of the Bitmap that Scuba tests need to allocate for screen diffing often exceed what any given app on the device can allocate given heap size constraints.

Moreover, the images used in sample data for some SCUBA tests can be very large as well - when inlined into a proto, the ComponentView framework tries to allocate a Bitmap for them. This can also exceed current memory limits.

This defines a new emulator target, swordfish_scuba, that is otherwise identical to a swordfish device save for a greatly increased memory and heap limit. It also updates fruit's ANDROID_WEAR scuba test targets to use this emulator.

This emulator should *only* be used for SCUBA tests, since it is an unrealistic amount of memory for current generation Wear devices. Tests that target this emulator may not catch bugs that would result in OOM errors on real-world devices.

PiperOrigin-RevId: 181605598