Split dev repositories from the unified launcher repositories.
This prevents introducing unnecessary external repositories to the top level namespace of the user's Bazel project. e.g. the user already has `@guava` defined, and loading `repo.bzl` would cause a conflict. This CL solves that.

Since the test entry point is a deploy JAR, there is no need for the users to
download the external dependencies for it. We only need to do this for the Unified Launcher's deps, which are already namespaced with "_archive".

PiperOrigin-RevId: 188887505