Stabilizes moveToState() and launch() implementation on device.
There are three problems in moveToState() and one problem in launch() in the previous implementation.

Problems in moveToState():

1) ActivityScenario does not close helper activities (EmptyActivity, EmptyFloatingActivity) after each test case runs, it potentially causes disruption for subsequent tests.

This CL changes InstrumentationActivityInvoker#startActivity to close these helper activities before starting tested activity. Also, now #moveToState() accepts State.DESTROY and you can call moveToState(DESTROYED) from your test to clean up the state explicitly.

2) Race-condition between intents to start Empty, EmptyFloatingActivity and the timing of its destruction.

InstrumentationActivityInvoker controls the lifecycle by starting activities with Intent.FLAG_ACTIVITY_REORDER_TO_FRONT flag. It turns out there is a race condition between intents to start activities and the timing of its destruction. This CL changes the strategy to communicate EmptyActivity and the invoker via broadcast. Instead of re-starting activities with the flag, send FINISH_ACTIVITY action intent to EmptyActivity to finish itself to bring back the activity under test to be the top of activity stack.

3) ActivityScenario didn't filter out incoming lifecycle events callback for those old activities.

Add more filters in ActivityLifecycleObserver in ActivityScenario not to confused by the previous instance.

A problem in launch():

When you call #finish in your Activity#onCreate(), a specialized lifecycle transition is applied and #onResume() is never called. This CL changes it to wait for Activity to become either resumed or finished.

PiperOrigin-RevId: 217917039