Implement getResult() method in ActivityScenario where you can retrieve the target activity's result code and data.
As opposed to the easiness of the feature request description, implementing it is hard because there is no public API to access to the result code and data directly from your activity or instrumentation. So, I implemented this functionality by having BootstrapActivity. From the instrumentation thread, you create PendingIntent which starts your target activity and passes it to BootstrapActivity. BootstrapActivity uses the PendingIntent and starts it by startIntentSenderForResult. After the target activity finishes, BootstrapActivity#onActivityResult is called and it forwards the data back to the instrumentation thread via broadcast. This is because BootstrapActivity may be declared in instrumentation APK so it may run in the different process. Finally, ActivityResultWaiter listens to the broadcast and publish the data to the ActivityScenario. Please see InstrumentationActivityInvoker for more details.

PiperOrigin-RevId: 221313151