Replaced the deprecated #findViewHolderForPosition method with its replacement #findViewHolderForAdapterPosition.
The other option is to replace with #findViewHolderForLayoutPosition, whose implementation is identical to #findViewHolderForPosition's. However, since #findViewHolderForAdapterPosition is called in the #onLayout method, using #findViewHolderForAdapterPosition will have an advantage when users use proguard in their tests. In addition, the differences between #findViewHolderForLayoutPosition and #findViewHolderForAdapterPosition is that the latter takes into account "any pending adapter changes that may not be reflected to the layout yet". In Espresso, we should sync before operating on the RecyclerView anyway, so these two methods probably shouldn't have a difference in our context.

https://github.com/android/android-test/issues/354

PiperOrigin-RevId: 251244591