Attempt to reduce the log noise in AXT Google Cloud Build
The log size of the AXT Google Cloud Build is massive, making it tedious to
find the actual error messages. Most of the output is the result of progress
meters when downloading and installing components. Add options to quiet the
output of some of the commands while still retaining error messages.

PiperOrigin-RevId: 297429271