Streamline TestLoader implementation and error handling.
TestLoader is responsible for loading candidate classes,
and determining if they are test classes.

It essentially has two operation modes:
a. Class path scanning mode. This mode is scanning through the classpath
looking for tests. Error handling in this mode should be more lenient, as
its expected that not all classes encountered are actually runnable tests.
2. Direct mode. In this mode the caller has directly specified the test class
names to run, and if any class specified is not a runnable test an error should
be reported.

There are three main problems being addressed with this change:
1. TestLoader was not consistently reporting errors in direct mode. For example
   if a class could not be found at all, it would return an 'UnloadableClassloader' that
   contained a Description with no children. However if the class could be loaded, but it
   failed during JUnit initialization, a Description with a single 'initializationError'
   child would be created.
2. Classpath scanning vs direct mode was essentially implemented via the 'Flag Argument' anitpattern
    https://martinfowler.com/bliki/FlagArgument.html. Even worse, the flag was passed into both builders
    and the TestLoader, leaking abstraction details.
3. The TestLoader unit tests tested implementation details - focusing on the specific error classes being
   returned rather than the behavior when executed.

This change addressed these problems by:
a. Creating specific subclasses for TestLoader for each behavior type
b. Removing knowledge of the different loader modes from the Builder
c. Applying a ErrorRunner consistent with upstream JUnit if any loading failures in direct mode
d. Cleanup the unit tests - moving inner fixture classes to the top level, and changing logic to
test behavior,

PiperOrigin-RevId: 407474125