Add a ViewAction that scrolls to the last position in a RecyclerView.
Note that we are scrolling to the last position instead of the bottom or end. This is because the last element may not be at the bottom/end as this depends on the LayoutManager e.g. LinearLayoutManager has a setReverseLayout method that puts the first element at the bottom/end: https://developer.android.com/reference/kotlin/androidx/recyclerview/widget/LinearLayoutManager#setreverselayout.

PiperOrigin-RevId: 408480886