Truncate view hierarchy in exception messages when it gets too large.
The view matcher exceptions dump the view hierarchy as well as the stack trace where the view matching fail. When the view hierarchy is too large, that stack trace is missing in the output.
This truncates the view hierarchy in the exception message to 64KB including 2KB room for the stack trace.

If a full view hierarchy dump file has been created, that information is also added to the output message.

This also adds a new Espresso sample to generate truncated view hierarchy, with a matching failing test to validate the output contains the view hierarchy and that it is truncated enough to contain the relevant stack trace.

PiperOrigin-RevId: 416328749